+++ config/src/ActivityDefinition.ts
@@ -1,0 +1,6 @@
+import type { ActivityLoader } from "./ActivityLoader";
+
+export interface ActivityDefinition<ActivityName extends string> {
+  name: ActivityName;
+  loader?: ActivityLoader<any>;
+}