+++ integrations/react/src/future/ActivityComponentType.tsx
@@ -1,0 +1,9 @@
+import type {
+  InferActivityParams,
+  RegisteredActivityParamTypes,
+} from "@stackflow/config";
+import type React from "react";
+
+export type ActivityComponentType<
+  ActivityName extends Extract<keyof RegisteredActivityParamTypes, string>,
+> = React.ComponentType<{ params: InferActivityParams<ActivityName> }>;