+++ integrations/react/src/future/StackComponentType.ts
@@ -1,0 +1,3 @@
+export type StackComponentType = React.FC<{
+  initialLoaderData?: any;
+}>;